🚚 Move week 01 problems to 공통주차별문제.Week01.MIRIM
